+++ opm/models/flash/flashproperties.hh
@@ -31,9 +31,9 @@
 #define EWOMS_FLASH_PROPERTIES_HH
 
 #include <opm/models/common/multiphasebaseproperties.hh>
-#include <ewoms/io/vtkcompositionmodule.hh>
-#include <ewoms/io/vtkenergymodule.hh>
-#include <ewoms/io/vtkdiffusionmodule.hh>
+#include <opm/models/io/vtkcompositionmodule.hh>
+#include <opm/models/io/vtkenergymodule.hh>
+#include <opm/models/io/vtkdiffusionmodule.hh>
 
 BEGIN_PROPERTIES
