+++ opm/models/immiscible/immiscibleproperties.hh
@@ -31,7 +31,7 @@
 #define EWOMS_IMMISCIBLE_PROPERTIES_HH
 
 #include <opm/models/common/multiphasebaseproperties.hh>
-#include <ewoms/io/vtkenergymodule.hh>
+#include <opm/models/io/vtkenergymodule.hh>
 
 BEGIN_PROPERTIES
