+++ opm/models/io/vtkprimaryvarsmodule.hh
@@ -27,8 +27,8 @@
 #ifndef EWOMS_VTK_PRIMARY_VARS_MODULE_HH
 #define EWOMS_VTK_PRIMARY_VARS_MODULE_HH
 
-#include <ewoms/io/baseoutputmodule.hh>
-#include <ewoms/io/vtkmultiwriter.hh>
+#include <opm/models/io/baseoutputmodule.hh>
+#include <opm/models/io/vtkmultiwriter.hh>
 
 #include <opm/models/utils/parametersystem.hh>
 #include <opm/models/utils/propertysystem.hh>
