+++ opm/models/io/vtkscalarfunction.hh
@@ -22,16 +22,19 @@
 */
 /*!
  * \file
- * \copydoc Opm::VtkTensorFunction
+ * \copydoc Opm::VtkScalarFunction
  */
-#ifndef VTK_TENSOR_FUNCTION_HH
-#define VTK_TENSOR_FUNCTION_HH
+#ifndef VTK_SCALAR_FUNCTION_HH
+#define VTK_SCALAR_FUNCTION_HH
 
-#include <ewoms/io/baseoutputwriter.hh>
+#include <opm/models/io/baseoutputwriter.hh>
 
 #include <dune/grid/io/file/vtk/function.hh>
+#include <dune/istl/bvector.hh>
 #include <dune/common/fvector.hh>
 #include <dune/common/version.hh>
+
+#include <opm/material/common/Unused.hpp>
 
 #include <opm/material/common/Exceptions.hpp>
 
@@ -42,46 +45,45 @@
 namespace Opm {
 
 /*!
- * \brief Provides a tensor-valued function using Dune::FieldMatrix objects as elements.
+ * \brief Provides a vector-valued function using Dune::FieldVectors
+ *        as elements.
  */
 template <class GridView, class Mapper>
-class VtkTensorFunction : public Dune::VTKFunction<GridView>
+class VtkScalarFunction : public Dune::VTKFunction<GridView>
 {
     enum { dim = GridView::dimension };
     typedef typename GridView::ctype ctype;
     typedef typename GridView::template Codim<0>::Entity Element;
 
-    typedef BaseOutputWriter::TensorBuffer TensorBuffer;
+    typedef BaseOutputWriter::ScalarBuffer ScalarBuffer;
 
 public:
-    VtkTensorFunction(std::string name,
+    VtkScalarFunction(std::string name,
                       const GridView& gridView,
                       const Mapper& mapper,
-                      const TensorBuffer& buf,
-                      unsigned codim,
-                      unsigned matrixColumnIdx)
+                      const ScalarBuffer& buf,
+                      unsigned codim)
         : name_(name)
         , gridView_(gridView)
         , mapper_(mapper)
         , buf_(buf)
         , codim_(codim)
-        , matrixColumnIdx_(matrixColumnIdx)
     { assert(int(buf_.size()) == int(mapper_.size())); }
 
     virtual std::string name() const
     { return name_; }
 
     virtual int ncomps() const
-    { return static_cast<int>(buf_[0].M()); }
+    { return 1; }
 
-    virtual double evaluate(int mycomp,
+    virtual double evaluate(int mycomp OPM_UNUSED,
                             const Element& e,
                             const Dune::FieldVector<ctype, dim>& xi) const
     {
-        size_t idx;
+        unsigned idx;
         if (codim_ == 0) {
             // cells. map element to the index
-            idx = static_cast<size_t>(mapper_.index(e));
+            idx = static_cast<unsigned>(mapper_.index(e));
         }
         else if (codim_ == dim) {
             // find vertex which is closest to xi in local
@@ -97,29 +99,26 @@ public:
                 local -= xi;
                 if (local.infinity_norm() < min) {
                     min = local.infinity_norm();
-                    imin = i;
+                    imin = static_cast<int>(i);
                 }
             }
 
             // map vertex to an index
-            idx = static_cast<size_t>(mapper_.subIndex(e, imin, codim_));
+            idx = static_cast<unsigned>(mapper_.subIndex(e, imin, codim_));
         }
         else
-            throw std::logic_error("Only element and vertex based tensor fields are supported so far.");
+            throw std::logic_error("Only element and vertex based vector fields are"
+                                   " supported so far.");
 
-        unsigned i = static_cast<unsigned>(mycomp);
-        unsigned j = static_cast<unsigned>(matrixColumnIdx_);
-
-        return static_cast<double>(static_cast<float>(buf_[idx][i][j]));
+        return static_cast<double>(static_cast<float>(buf_[idx]));
     }
 
 private:
     const std::string name_;
     const GridView gridView_;
     const Mapper& mapper_;
-    const TensorBuffer& buf_;
+    const ScalarBuffer& buf_;
     unsigned codim_;
-    unsigned matrixColumnIdx_;
 };
 
 } // namespace Opm
