+++ opm/models/utils/basicproperties.hh
@@ -32,7 +32,7 @@
 
 #include <opm/models/utils/propertysystem.hh>
 #include <opm/models/utils/parametersystem.hh>
-#include <ewoms/io/dgfvanguard.hh>
+#include <opm/models/io/dgfvanguard.hh>
 
 #if HAVE_DUNE_FEM
 #include <dune/fem/gridpart/adaptiveleafgridpart.hh>
