+++ tests/problems/lensproblem.hh
@@ -28,7 +28,7 @@
 #ifndef EWOMS_LENS_PROBLEM_HH
 #define EWOMS_LENS_PROBLEM_HH
 
-#include <ewoms/io/structuredgridvanguard.hh>
+#include <opm/models/io/structuredgridvanguard.hh>
 #include <opm/models/immiscible/immiscibleproperties.hh>
 #include <opm/models/discretization/common/fvbaseadlocallinearizer.hh>
 #include <opm/models/discretization/ecfv/ecfvdiscretization.hh>
